[Refactor] Modify navigation structure
