+++ src/types/NavigationTypes.ts
@@ -1,0 +1,12 @@
+/** Main 화면별 params 타입 지정 */
+export type MainStackParamList = {
+  MainScreen: undefined;
+  ProfileViewScreen: undefined;
+  ArticleViewScreen: undefined;
+  FeedScreen: undefined;
+};
+
+/** Auth 화면별 params 타입 지정 */
+export type AuthStackParamList = {
+  LoginScreen: undefined;
+};
